refactoring + icon
