+++ src/lib/components/icons/CartIcon.svelte
